+++ src/border.h
@@ -1,13 +1,10 @@
 #ifndef BORDER_H
 #define BORDER_H
 
-#define kCGSIgnoreForExposeTagBit (1 << 7)
-#define kCGSIgnoreForEventsTagBit (1 << 9)
-#define kCGSDisableShadowTagBit   (1 << 3)
-
 struct border
 {
     uint32_t id;
+    CFArrayRef id_ref;
     CGContextRef context;
     CFTypeRef region;
     CGRect frame;
@@ -15,13 +12,16 @@ struct border
 };
 
 struct window;
+bool border_should_order_in(struct window *window);
+void border_show_all(void);
+void border_hide_all(void);
+void border_hide(struct window *window);
+void border_show(struct window *window);
+void border_ensure_same_space(struct window *window);
 void border_redraw(struct window *window);
 void border_activate(struct window *window);
 void border_deactivate(struct window *window);
-void border_enter_fullscreen(struct window *window);
-void border_exit_fullscreen(struct window *window);
 void border_create(struct window *window);
-void border_resize(struct window *window);
 void border_destroy(struct window *window);
 
 #endif