+++ src/workspace.h
@@ -17,7 +17,7 @@ bool workspace_application_is_observable(struct process *process);
 bool workspace_application_is_finished_launching(struct process *process);
 void workspace_application_observe_finished_launching(void *context, struct process *process);
 void workspace_application_observe_activation_policy(void *context, struct process *process);
-int workspace_display_has_notch(uint32_t did);
+int workspace_display_notch_height(uint32_t did);
 pid_t workspace_get_dock_pid(void);
 bool workspace_is_macos_monterey(void);
 bool workspace_is_macos_bigsur(void);